% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{Servidor de la app}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Objetos que provee Shiny.}
}
\description{
Servidor de la app
}
\keyword{internal}
