% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Construye la interfaz de la app}
\usage{
app_ui()
}
\value{
Un \code{shiny.tag.list} con la interfaz completa.
}
\description{
Construye la interfaz de la app
}
\keyword{internal}
