% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diariosuy-package.R
\docType{package}
\name{diariosuy-package}
\alias{diariosuy}
\alias{diariosuy-package}
\title{diariosuy: Buscador y Analizador de Diarios de Sesión del Parlamento Uruguayo}
\description{
Busca sesiones parlamentarias por palabra clave en el sitio del Parlamento del Uruguay, descarga los diarios de sesión en PDF y analiza sus textos: frecuencia de términos, evolución en el tiempo y co-ocurrencias. La app se ejecuta en la máquina del usuario, condición necesaria para consultar el sitio oficial, que rechaza las conexiones desde fuera de Uruguay.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilianoGonLas/diariosuy}
  \item Report bugs at \url{https://github.com/EmilianoGonLas/diariosuy/issues}
}

}
\author{
\strong{Maintainer}: Emiliano González \email{emiliano.gonzalez.lasa@fing.edu.uy}

}
\keyword{internal}
