% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizar.R
\name{grafico_coocurrencias}
\alias{grafico_coocurrencias}
\title{Gráfico de co-ocurrencias entre el término principal y términos secundarios}
\usage{
grafico_coocurrencias(datos, terminos_cruce)
}
\arguments{
\item{datos}{tibble resultado de analizar_pdfs()}

\item{terminos_cruce}{lista nombrada usada en el análisis}
}
\value{
objeto plotly o NULL si no hay términos de cruce
}
\description{
Gráfico de co-ocurrencias entre el término principal y términos secundarios
}
