% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizar.R
\name{grafico_timeline}
\alias{grafico_timeline}
\title{Gráfico interactivo de menciones a lo largo del tiempo}
\usage{
grafico_timeline(datos, titulo = "Menciones del término en el tiempo")
}
\arguments{
\item{datos}{tibble resultado de analizar_pdfs()}

\item{titulo}{Título del gráfico}
}
\value{
objeto plotly
}
\description{
Gráfico interactivo de menciones a lo largo del tiempo
}
